other entities

+++ src/main/java/com/pi/dao/DepartmentDao.java
@@ -33,7 +33,7 @@ public class DepartmentDao {
         return entityManager.createQuery("SELECT d FROM Department d", Department.class).getResultList();
     }
 
-    public Department getDepartmentById(int id) {
+    public Department getDepartmentById(long id) {
         return entityManager.find(Department.class, id);
     }
